Edit values after update country.

+++ app/src/main/java/com/example/currencyconverter/domain/model/Country.kt
@@ -6,6 +6,6 @@ data class Country(
     val name: String,
     val currency: String,
     val currencyName: String,
-    val sendingLimit: Int,
+    val sendingLimit: Float,
     @DrawableRes val icon: Int
 )
